+++ docs/Changelog.md
@@ -7,6 +7,7 @@
 - POI Search [#728](https://github.com/mapsforge/mapsforge/issues/728)
 - Reverse Geocoding example [#727](https://github.com/mapsforge/mapsforge/issues/727)
 - LabelLayer improved implementation [#763](https://github.com/mapsforge/mapsforge/issues/763)
+- Night mode map rendering [#777](https://github.com/mapsforge/mapsforge/issues/777)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aissue+is%3Aclosed+milestone%3A0.6.1)